FEEL Visitors: Move visitors in separate package
