+++ dmn-core/src/main/java/com/gs/dmn/feel/synthesis/ReplaceItemFilterVisitor.java
@@ -10,12 +10,12 @@
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package com.gs.dmn.feel.analysis.semantics;
+package com.gs.dmn.feel.synthesis;
 
 import com.gs.dmn.error.ErrorHandler;
-import com.gs.dmn.feel.analysis.syntax.ast.CloneVisitor;
 import com.gs.dmn.feel.analysis.syntax.ast.Element;
 import com.gs.dmn.feel.analysis.syntax.ast.expression.Name;
+import com.gs.dmn.feel.analysis.syntax.ast.visitor.CloneVisitor;
 
 public class ReplaceItemFilterVisitor<T, C> extends CloneVisitor<T, C> {
     private final String oldLambdaParameterName;
