+++ dmn-core/src/test/java/com/gs/dmn/feel/analysis/syntax/ast/visitor/NamesVisitorTest.java
@@ -1,4 +1,4 @@
-package com.gs.dmn.feel.analysis.semantics;
+package com.gs.dmn.feel.analysis.syntax.ast.visitor;
 
 import com.gs.dmn.context.DMNContext;
 import com.gs.dmn.el.analysis.semantics.type.Type;
